remove SyncManager
