+++ curp/src/lib.rs
@@ -130,11 +130,11 @@ mod message;
 /// Log
 mod log;
 
-/// The special keybased channels
+/// The special keyBased channels
 mod channel;
 
-/// Sync manager to sync command to all other peers
-mod sync_manager;
+/// Background tasks of Curp protocol
+mod bg_tasks;
 
 /// Protobuf generated types that are used in RPC
 mod rpc;
